Lock ftp pod to carmd-er-lxtbd4bed9b5df59
It seems that ftp passive requires the IP to point to the pod directly,
and going through a service doesn't work well.
For this reason, IP of the node (10.1.60.147) has been advertised as the
IP to reach the ftp server.

Since this IP is associated with the node, it could change if the pod
is rescheduled on some other node (if it goes down or has some resource
constrain).

As to make sure sure that the 10.1.60.147 will continue serving the ftp
service, updating the resource to add a constrain.

The rollingUpdate configuration is there as by default it would try to
create another pod on the same machine, and that's not possible as the
pod is using host ports. The rollingUpdate configuration defines that
there would be only one pod running at the same time.

Resolves: LETEST-5803
Change-Id: Id6cf300220f8370266aee4811e6c902f0b76b67b